% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{diff_geo_means}
\alias{diff_geo_means}
\title{diff_geo_means}
\usage{
diff_geo_means(data, i, percentage = FALSE)
}
\arguments{
\item{data}{.}

\item{i}{.}

\item{percentage}{Return the difference as a percentage (default = `FALSE`)}
}
\value{

}
\description{
diff_geo_means
}
