% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{log_ratio_difference}
\alias{log_ratio_difference}
\title{log_ratio_difference}
\usage{
log_ratio_difference(
  composition,
  group,
  n_rep = 1000,
  alpha = 0.05,
  confint_type = "bca"
)
}
\arguments{
\item{composition}{A composition (from the acomp function)}

\item{group}{A vector (single column) of the grouping variable.
This variable must only have two levels (e.g. 'Male' and 'Female').}

\item{n_rep}{Number of bootstrap iterations}

\item{alpha}{Alpha (i.e. p value) to use for confidence intervals}

\item{confint_type}{Type of confidence interval. One of ("norm","basic", "stud", "perc", "bca").
Default value is 'bca'}
}
\value{

}
\description{
Returns the log-ratio difference for each component of a composition
between two groups. The confidence intervals of the difference are estimated
via bootstrapping.
}
\examples{

}
\references{
Martín Fernández, Josep Antoni, Josep Daunis i Estadella, and Glòria Mateu i Figueras.
"On the interpretation of differences between groups for compositional data."
SORT: statistics and operations research transactions, 2015, vol. 39, núm. 2, p. 231-252 (2015).
}
