% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pairwise_hotelling_test}
\alias{pairwise_hotelling_test}
\title{pairwise_hotelling_test}
\usage{
pairwise_hotelling_test(comp, groups, adjust = c("holm"))
}
\arguments{
\item{comp}{A dataframe or composition object of multiple parts}

\item{groups}{A vector of the grouping factor. Must be same length as `comp`}

\item{adjust}{Method of p value adjustment (default ='holm'). See \code{\link[stats]{p.adjust.methods}}}
}
\value{
A dataframe of pairwise contrasts
}
\description{
The Hotelling T-squared test is the multivariate equivalent of a t-test. This
function can be used as a post-hoc test to a MANOVA.
}
\examples{

}
