% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_geo_means}
\alias{plot_geo_means}
\title{plot_geo_means}
\usage{
plot_geo_means(composition, group, type = "both")
}
\arguments{
\item{composition}{A composition (from the acomp function)}

\item{group}{A vector (single column) of the grouping variable. This can be
any number of levels.}

\item{type}{The type of plot to return. Can be one of:
\itemize{
  \item "component" The colours of the bars represent the components of the composition
  \item "group" The colours of the bars represent the levels of `group`.
  \item "both" (default) Returns a 2-panel plot with both 'component' and 'group' plots.
}}
}
\value{

}
\description{
Creates a geometric mean barplot to visualise differences between
groups. The logratio between the whole geometric mean and the geometric mean of
the group is calculated. These are presented on a log scale.
}
\examples{

}
\references{
Martín Fernández, Josep Antoni, Josep Daunis i Estadella, and Glòria Mateu i Figueras.
"On the interpretation of differences between groups for compositional data."
SORT: statistics and operations research transactions, 2015, vol. 39, núm. 2, p. 231-252 (2015).
}
