% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_log_ratio_difference}
\alias{plot_log_ratio_difference}
\title{plot_log_ratio_difference}
\usage{
plot_log_ratio_difference(data, h_line_color = "black")
}
\arguments{
\item{data}{A dataframe output from \code{\link{log_ratio_difference}}}

\item{h_line_color}{String. Color of the horozontal line crossing 0.}
}
\value{

}
\description{
Plots the log-ratio difference and confidence intervals that
were estimated from the \code{\link{log_ratio_difference}} function.
}
\examples{

}
\references{
Martín Fernández, Josep Antoni, Josep Daunis i Estadella, and Glòria Mateu i Figueras.
"On the interpretation of differences between groups for compositional data."
SORT: statistics and operations research transactions, 2015, vol. 39, núm. 2, p. 231-252 (2015).
}
