% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_geo_means}
\alias{plot_geo_means}
\title{plot_geo_means}
\usage{
plot_geo_means(composition, group)
}
\arguments{
\item{composition}{A composition (from the acomp function)}

\item{group}{A vector (single column) of the grouping variable. This can be
any number of levels.}
}
\value{

}
\description{
Creates a geometric mean barplot to visualise differences between
groups.
}
\examples{

}
\references{
Martín Fernández, Josep Antoni, Josep Daunis i Estadella, and Glòria Mateu i Figueras.
"On the interpretation of differences between groups for compositional data."
SORT: statistics and operations research transactions, 2015, vol. 39, núm. 2, p. 231-252 (2015).
}
